Updated unit test case for AssignedAgentData
